requirements.txt
chess==1.9.4
neat-python==0.92
numpy~=1.24.3
chess-board==0.3.1